# docs/openapi.json
{
  "components": {
    "schemas": {
      "Body_login_api_v1_login_post": {
        "properties": {
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          },
          "grant_type": {
            "anyOf": [
              {
                "pattern": "^password$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "password": {
            "format": "password",
            "title": "Password",
            "type": "string"
          },
          "scope": {
            "default": "",
            "title": "Scope",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_api_v1_login_post",
        "type": "object"
      },
...
      "Detail": {
        "description": "FastAPI's error body, written down so it can be named in `responses=`.\n\nEvery non-validation error this API returns is one of these. Declaring it\nis what turns \"401\" in the docs from a bare number into a shape a client\ncan code against.",
        "example": {
          "detail": "Could not validate credentials"
        },
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Detail",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Token": {
        "description": "What `/login` and `/auth/refresh` hand back.\n\nNote what *isn't* here: the refresh token. It goes out in an httpOnly\ncookie, where no script on the page can read it, which is the whole reason\nthere are two tokens rather than one long-lived one.",
        "example": {
          "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjo3fQ.sig",
          "csrf_token": "0lNrK8s2Qm6Xv3dYpHtW1cJfB9aZuE4g",
          "expires_in": 900,
          "token_type": "bearer"
        },
        "properties": {
          "access_token": {
            "description": "Send as `Authorization: Bearer …`. Hold it in memory, not in storage — it expires in minutes and `/auth/refresh` replaces it.",
            "title": "Access Token",
            "type": "string"
          },
          "csrf_token": {
            "description": "Echo this back as an `X-CSRF-Token` header on `/auth/refresh` and `/auth/logout`. It changes on every rotation.",
            "title": "Csrf Token",
            "type": "string"
          },
          "expires_in": {
            "description": "Seconds until the access token expires. Refresh a little before this rather than waiting for a 401.",
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "csrf_token"
        ],
        "title": "Token",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Source on GitHub",
      "url": "https://github.com/AndrewTechTips/social-feed-application"
    },
    "description": "The backend behind **Commons** — a small public feed.\n\nPosts, comments, upvotes and accounts, with a two-token auth flow. Reading is public; writing needs an account.\n\nThe published demo at [andrewtechtips.github.io/social-feed-application](https://andrewtechtips.github.io/social-feed-application/) reimplements this same contract in the browser, because a static host has nowhere to run FastAPI. This document is the contract both of them answer to.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/AndrewTechTips/social-feed-application/blob/main/LICENSE"
    },
    "title": "Social Feed API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/logout-all": {
      "post": {
        "description": "End every session this account has, including this one.\n\nAuthenticated with the *access* token rather than the refresh cookie, which\nis the difference between this and `/logout`. Signing out one browser is\nsomething a browser asks for and needs no proof beyond holding the cookie;\nsigning out every browser is something a person asks for, and it should\ntake more than having once been handed a cookie to do it.\n\nThat is also why there is a `refresh_sessions` table at all. A stateless\nrefresh token cannot be taken back, so this would have been a button that\ncleared one cookie and lied about the rest.",
        "operationId": "logout_everywhere_api_v1_auth_logout_all_post",
        "responses": {
          "204": {
            "description": "Every session ended, on every machine."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No access token, or one that has expired, been tampered with, or belongs to a deleted account. Trade the refresh cookie for a new token at `POST /auth/refresh` rather than sending the person back to a sign-in form."
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Sign out everywhere",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Rotate the session.\n\nEvery call invalidates the token that was presented and issues the next\none, so a copied cookie is detectable: the copy eventually presents a\nsecret that has already been rotated away, and the whole session is ended\nrather than guessing which of the two holders is the thief.\n\nSend the `csrf_token` from the last `/login` or `/auth/refresh` response in\nan `X-CSRF-Token` header. Without it this answers 401 even with a perfectly\ngood cookie.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No usable refresh cookie, or no matching `X-CSRF-Token`. One answer for every way of being wrong — saying which would tell a caller whether they hold half of a real session. The cookie is cleared, because whatever is in the browser is no longer worth sending."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "Too many attempts (30/minute)."
          }
        },
        "summary": "Trade the refresh cookie for a new access token",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/login": {
      "post": {
        "description": "Exchange an email and a password for an access token and a refresh cookie.\n\nThe form field is called `username` because that is what the OAuth2\npassword flow calls it, and `/docs` builds its form from the same spec.\nThis API signs people in by **email** — the username is the public identity\nand is not a credential.",
        "operationId": "login_api_v1_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_api_v1_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "The email and password don't match. Deliberately the same answer whether or not the address has an account."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "Too many attempts (5/minute)."
          }
        },
        "summary": "Sign in",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "A local development server",
      "url": "http://localhost:8000"
    }
  ],
...
}